feat: add description preprocessing for search items
This commit introduces the ability to preprocess the description of search items. A new model, ItemDescriptionPreprocess, has been created to store preprocessing configurations. The Item model is updated to use this preprocessing when setting the description. The bulk items creator also now ensures the description hash is set when creating items. The language model query function has been updated to include files in its parameters.